Version funcional con opciones para subir archivos en Single y Bulk. Profiling incluido
